fix(linear): share oauth state helpers
